+++ src/main/java/sparta/streaming/domain/user/CustomOAuth2User.java
@@ -1,4 +1,4 @@
-package sparta.streaming.domain;
+package sparta.streaming.domain.user;
 
 
 import lombok.AllArgsConstructor;
